% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxa.R
\name{update_taxa}
\alias{update_taxa}
\title{Update species taxa based on speciesmaster lookup}
\usage{
update_taxa(vegdata, sppmaster)
}
\arguments{
\item{vegdata}{is a data.table with species codes in a 'Species' column}

\item{sppmaster}{the taxon lookup table from the spp master}
}
\description{
Update species taxa based on speciesmaster lookup
}
